Move Token out of literal.h; Add DEFINE_NAMED_TOKEN
